Make the functionality in one file
