+++ frontend/hub/collections/UploadCollection.tsx
@@ -1,7 +1,15 @@
 import { Alert, Radio } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import { PageDetails, PageForm, PageHeader, PageLayout } from '../../../framework';
+import {
+  PageDetails,
+  PageForm,
+  PageHeader,
+  PageLayout,
+  PageTable,
+  TextCell,
+  ITableColumn,
+} from '../../../framework';
 import { PageDetail } from '../../../framework/PageDetails/PageDetail';
 import { PageFormFileUpload } from '../../../framework/PageForm/Inputs/PageFormFileUpload';
 import { PageFormWatch } from '../../../framework/PageForm/Utils/PageFormWatch';
@@ -11,11 +19,24 @@ import { postRequestFile } from '../../common/crud/Data';
 import { useHubNamespaces } from '../namespaces/hooks/useHubNamespaces';
 import { useRepositories } from '../repositories/hooks/useRepositories';
 import { hubAPI } from '../api';
-import { useState } from 'react';
-import { MultipleRepoSelector } from './MultipleRepoSelector';
+
+import { useState, useMemo } from 'react';
+
+import { usePulpView } from '../usePulpView';
+import { IToolbarFilter } from '../../../framework';
+import { nameKeyFn } from '../api';
 
 interface UploadData {
   file: unknown;
+}
+
+export interface Repository {
+  name: string;
+  description?: string;
+  pulp_id: string;
+  pulp_last_updated: string;
+  content_count: number;
+  gpgkey: string | null;
 }
 
 export function UploadCollection() {
@@ -40,7 +61,20 @@ export function UploadCollectionByFile() {
   const repositories = useRepositories();
   const navigate = useNavigate();
   const onCancel = () => navigate(-1);
+
+  const toolbarFilters = useRepoFilters();
+
+  // const navigate = useNavigate()
+  const tableColumns = useRepositoriesColumns();
+  const view = usePulpView<Repository>({
+    url: '/api/automation-hub/pulp/api/v3/repositories/ansible/ansible/',
+    keyFn: nameKeyFn,
+    toolbarFilters,
+    tableColumns,
+  });
+
   const [onlyStaging, setOnlyStaging] = useState(true);
+  //const [selectedItems, setSelectedItems] = useState<string[]>([]);
 
   function renderRepoSelector() {
     return (
@@ -68,7 +102,21 @@ export function UploadCollectionByFile() {
             <>{t`Please note that those repositories are not filtered by permission, if operation fail, you don't have it.`}</>
           )}
         </div>
-        <MultipleRepoSelector />
+
+        <PageTable<Repository>
+          isSelectMultiple={false}
+          showSelect={true}
+          disableColumnManagement={true}
+          disableListView={true}
+          disableCardView={true}
+          tableColumns={tableColumns}
+          toolbarFilters={toolbarFilters}
+          errorStateTitle={t('Error loading repositories')}
+          emptyStateTitle={t('No repositories yet')}
+          emptyStateDescription={t('To get started, create an repository.')}
+          defaultTableView="table"
+          {...view}
+        />
       </>
     );
   }
@@ -121,5 +169,48 @@ export function UploadCollectionByFile() {
         </PageForm>
       )}
     </>
+  );
+}
+
+export function useRepositoriesColumns() {
+  const { t } = useTranslation();
+  const tableColumns = useMemo<ITableColumn<Repository>[]>(
+    () => [
+      {
+        header: t('Name'),
+        sort: 'name',
+        cell: (repository) => (
+          <>
+            <TextCell text={repository.name} />
+          </>
+        ),
+      },
+      {
+        header: t('Description'),
+        cell: (repository) => (
+          <>
+            <TextCell text={repository.description} />
+          </>
+        ),
+      },
+    ],
+    []
+  );
+  return tableColumns;
+}
+
+export function useRepoFilters() {
+  const { t } = useTranslation();
+  return useMemo<IToolbarFilter[]>(
+    () => [
+      {
+        key: 'name',
+        label: t('Name'),
+        type: 'string',
+        query: 'name',
+        comparison: 'startsWith',
+      },
+    ],
+    [t]
   );
 }
